feat: update package.json, remove dependency, and delete react-table
