+++ package.json
@@ -99,7 +99,6 @@
     "react-select": "^5.7.3",
     "react-slider": "^2.0.4",
     "react-spinners": "^0.13.8",
-    "react-table": "^7.8.0",
     "react-toastify": "^9.1.3",
     "react-tooltip": "^5.13.1",
     "reflect-metadata": "^0.1.13",
@@ -147,7 +146,6 @@
     "@types/react-grid-layout": "^1.3.2",
     "@types/react-qr-reader": "^2.1.4",
     "@types/react-slider": "^1.3.1",
-    "@types/react-table": "^7.7.14",
     "@types/secp256k1": "^4.0.3",
     "@types/styled-components": "^5.1.26",
     "@types/styled-react-modal": "^1.2.2",
